Moved layout/module logic from `Workbench`to new class
