+++ frontend/src/framework/internal/components/Content/private-components/layout.tsx
@@ -1,7 +1,8 @@
 import React from "react";
 
 import { GuiEvent, GuiEventPayloads } from "@framework/GuiMessageBroker";
-import { LayoutElement, Workbench } from "@framework/Workbench";
+import { LayoutElement } from "@framework/LayoutService";
+import { Workbench } from "@framework/Workbench";
 import { useModuleInstances } from "@framework/internal/hooks/workbenchHooks";
 import { useElementSize } from "@lib/hooks/useElementSize";
 import {
@@ -37,8 +38,9 @@ export const Layout: React.FC<LayoutProps> = (props) => {
     const mainRef = React.useRef<HTMLDivElement>(null);
     const size = useElementSize(ref);
     const layoutBoxRef = React.useRef<LayoutBox | null>(null);
-    const moduleInstances = useModuleInstances(props.workbench);
+    const moduleInstances = useModuleInstances(props.workbench.getLayoutService());
     const guiMessageBroker = props.workbench.getGuiMessageBroker();
+    const layoutService = props.workbench.getLayoutService();
 
     const convertLayoutRectToRealRect = React.useCallback(
         (element: LayoutElement): Rect => {
@@ -61,9 +63,9 @@ export const Layout: React.FC<LayoutProps> = (props) => {
         let dragging = false;
         let moduleInstanceId: string | null = null;
         let moduleName: string | null = null;
-        setLayout(props.workbench.getLayout());
-        let originalLayout: LayoutElement[] = props.workbench.getLayout();
-        let currentLayout: LayoutElement[] = props.workbench.getLayout();
+        setLayout(layoutService.getLayout());
+        let originalLayout: LayoutElement[] = layoutService.getLayout();
+        let currentLayout: LayoutElement[] = layoutService.getLayout();
         let originalLayoutBox = makeLayoutBoxes(originalLayout);
         let currentLayoutBox = originalLayoutBox;
         layoutBoxRef.current = currentLayoutBox;
@@ -119,7 +121,7 @@ export const Layout: React.FC<LayoutProps> = (props) => {
                 if (isNewModule && moduleName) {
                     const layoutElement = currentLayout.find((el) => el.moduleInstanceId === pointerDownElementId);
                     if (layoutElement) {
-                        const instance = props.workbench.makeAndAddModuleInstance(moduleName, layoutElement);
+                        const instance = layoutService.makeAndAddModuleInstance(moduleName, layoutElement);
                         layoutElement.moduleInstanceId = instance.getId();
                         layoutElement.moduleName = instance.getName();
                     }
@@ -132,7 +134,7 @@ export const Layout: React.FC<LayoutProps> = (props) => {
                 originalLayoutBox = currentLayoutBox;
                 layoutBoxRef.current = currentLayoutBox;
                 setLayout(currentLayout);
-                props.workbench.setLayout(currentLayout);
+                layoutService.setLayout(currentLayout);
                 setPosition({ x: 0, y: 0 });
                 setPointer({ x: -1, y: -1 });
                 e.stopPropagation();
@@ -228,13 +230,13 @@ export const Layout: React.FC<LayoutProps> = (props) => {
             if (dragging) {
                 return;
             }
-            props.workbench.removeModuleInstance(e.moduleInstanceId);
+            layoutService.removeModuleInstance(e.moduleInstanceId);
             currentLayoutBox.removeLayoutElement(e.moduleInstanceId);
             currentLayout = currentLayoutBox.toLayout();
             setLayout(currentLayout);
             originalLayout = currentLayout;
             originalLayoutBox = currentLayoutBox;
-            props.workbench.setLayout(currentLayout);
+            layoutService.setLayout(currentLayout);
         };
 
         const removeModuleHeaderPointerDownSubscriber = guiMessageBroker.subscribeToEvent(
