+++ frontend/src/framework/internal/hooks/workbenchHooks.ts
@@ -1,18 +1,18 @@
 import React from "react";
 
+import { LayoutEvents, LayoutService } from "@framework/LayoutService";
 import { ModuleInstance } from "@framework/ModuleInstance";
-import { Workbench, WorkbenchEvents } from "@framework/Workbench";
 
-export function useModuleInstances(workbench: Workbench): ModuleInstance<any>[] {
+export function useModuleInstances(layoutService: LayoutService): ModuleInstance<any>[] {
     const [moduleInstances, setModuleInstances] = React.useState<ModuleInstance<any>[]>([]);
 
     React.useEffect(() => {
         function handleModuleInstancesChange() {
-            setModuleInstances(workbench.getModuleInstances());
+            setModuleInstances(layoutService.getModuleInstances());
         }
 
-        const unsubscribeFunc = workbench.subscribe(
-            WorkbenchEvents.ModuleInstancesChanged,
+        const unsubscribeFunc = layoutService.subscribe(
+            LayoutEvents.ModuleInstancesChanged,
             handleModuleInstancesChange
         );
